requirements.txt
# For Apple Silicon (M1/M2/M3), install with:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/metal

torch==2.3.0
torchvision==0.18.0
pillow
ipywidgets
numpy